+++ src/app/component/delta-stint/delta-stint.component.ts
@@ -4,12 +4,16 @@ import { DeltaStintService } from "../../service/delta-stint.service";
 import { DeltaStint } from "../../model/delta-stint";
 import { StintService } from "../../service/stint.service";
 import { millisecondsToTimeString } from "../../util/date.util";
+import { TooltipDirective } from "../ui/tooltip/tooltip.directive";
+import { Tooltip } from "../ui/tooltip/tooltip";
+import { TooltipPosition } from '../ui/tooltip/tooltip.enums';
 
 @Component({
   selector: 'app-delta-stint',
   imports: [
     NgForOf,
-    NgClass
+    NgClass,
+    TooltipDirective
   ],
   templateUrl: './delta-stint.component.html',
   styleUrl: './delta-stint.component.css'
@@ -20,6 +24,8 @@ export class DeltaStintComponent {
 
   readonly deltaStints: Signal<DeltaStint[]> = this.deltaStintService.deltaStints;
   readonly totalDeltaMilliseconds: Signal<number> = this.deltaStintService.totalDeltaMilliseconds;
+  readonly deltaMilliseconds: Signal<number> = this.deltaStintService.deltaMilliseconds;
+
   readonly activeStintId = computed(() => {
     const activeStint = this.stintService.activeStint();
     if (activeStint) {
@@ -29,6 +35,13 @@ export class DeltaStintComponent {
   })
 
   isOpen: boolean = true;
+
+  protected tooltipInfo: Tooltip = {
+    footer: "",
+    paragraphs: ["Only completed stint"],
+    title: ""
+  }
+  protected tooltipPosition: TooltipPosition = TooltipPosition.ABOVE;
 
   getMillisecondsToTimeString(milliseconds: number): string {
     return millisecondsToTimeString(Math.abs(milliseconds));